refactor(model): deep cleanup — extract hydrateRelations, remove dead code, fix imports
- Extract hydrateRelations (~250 lines) from Ad4mModel.ts to hydration.ts
- Create query-common.ts with resolveParentPredicate (shared by prolog + surreal)
- Move buildConformanceFilter back to decorators.ts (was conformance.ts, caused circular dep)
- Remove dead code from query-surreal.ts (buildSurrealWhereClause, buildSurrealSelectFields, etc.)
- Remove dead code from decorators.ts (instanceToSerializable, PerspectiveAction, addLink, hasLink)
- Remove formatSurrealValue wrapper and backward-compat re-exports from Ad4mModel.ts
- Fix import paths: PerspectiveClient/PerspectiveProxy import types from ./types
- Update barrel index.ts with query-common.ts
