+++ core/src/model/index.ts
@@ -10,6 +10,7 @@ export * from "./types";
 export * from "./Ad4mModel";
 export * from "./decorators";
 export * from "./Subject";
+export * from "./query-common";
 export * from "./query-prolog";
 export * from "./prolog-facts";
 export * from "./json-schema";
